Add CoverageInformation type and moved scores component in specific folder
