+++ opencti-platform/opencti-front/src/private/components/analyses/security_coverages/SecurityCoverage-types.ts
@@ -1,0 +1,4 @@
+export interface CoverageInformation {
+  readonly coverage_name: string;
+  readonly coverage_score: number;
+}
