+++ opencti-platform/opencti-front/src/private/components/analyses/security_coverages/security_coverage_scores/SecurityCoverageScores.tsx
@@ -5,11 +5,12 @@ import { Avatar, Tooltip } from '@mui/material';
 import { BugReportOutlined, HourglassEmpty, Person, ShieldOutlined, TrackChangesOutlined } from '@mui/icons-material';
 import Chart from '@components/common/charts/Chart';
 import { ApexOptions } from 'apexcharts';
-import { useFormatter } from '../../../../components/i18n';
-import { isEmptyField, isNotEmptyField } from '../../../../utils/utils';
-import { donutChartOptions } from '../../../../utils/Charts';
-import type { Theme } from '../../../../components/Theme';
-import { capitalizeFirstLetter } from '../../../../utils/String';
+import { useFormatter } from '../../../../../components/i18n';
+import { isEmptyField, isNotEmptyField } from '../../../../../utils/utils';
+import { donutChartOptions } from '../../../../../utils/Charts';
+import type { Theme } from '../../../../../components/Theme';
+import { capitalizeFirstLetter } from '../../../../../utils/String';
+import { CoverageInformation } from '../SecurityCoverage-types';
 
 const useStyles = makeStyles((theme: Theme) => ({
   charts: {
@@ -53,10 +54,7 @@ const useStyles = makeStyles((theme: Theme) => ({
 }));
 
 interface SecurityCoverageScoresProps {
-  coverage_information: ReadonlyArray<{
-    readonly coverage_name: string;
-    readonly coverage_score: number;
-  }> | null | undefined;
+  coverage_information?: ReadonlyArray<CoverageInformation> | null;
   variant?: 'header' | 'details' | 'matrix';
 }
 
@@ -116,7 +114,7 @@ const SecurityCoverageScores: FunctionComponent<SecurityCoverageScoresProps> = (
         <div className={classes.chartContainer} style={{ width: size, height: size }}>
           <div className={classes.chart}>
             <Chart options={options} series={series} type="donut" width={chartSize} height={chartSize} />
-            <Tooltip title={t_i18n('No executable tests are currently set for this entity, these can be set in OpenAEV')} placement="bottom">
+            <Tooltip title="Pending" placement="bottom">
               <Avatar className={classes.iconOverlay} sx={{ bgcolor: 'transparent', width: iconSize, height: iconSize }} style={{ top: iconPosition, left: iconPosition, fontSize: iconSize - 2 }}>
                 <span style={{ color: theme.palette.text?.primary }}><HourglassEmpty style={{ fontSize: iconSize }} /></span>
               </Avatar>
@@ -132,7 +130,6 @@ const SecurityCoverageScores: FunctionComponent<SecurityCoverageScoresProps> = (
           return (
             <div key={coverageResult.coverage_name} className={classes.chartContainer} style={{ width: size, height: size, padding: variant === 'matrix' ? 2 : 4 }}>
               <div className={classes.chart}>
-                <>test</>
                 <Chart options={options} series={series} type="donut" width={chartSize} height={chartSize} />
                 <Tooltip title={`${coverageResult.coverage_name} ${coverageResult.coverage_score}/100`} placement="bottom">
                   <Avatar className={classes.iconOverlay} sx={{ bgcolor: 'transparent', width: iconSize, height: iconSize }} style={{ top: iconPosition, left: iconPosition - 1, fontSize: iconSize - 2 }}>
@@ -156,7 +153,7 @@ const SecurityCoverageScores: FunctionComponent<SecurityCoverageScoresProps> = (
           <div className={classes.chartContainer}>
             <div className={classes.chart}>
               <Chart options={options} series={series} type="donut" width={70} height={70} />
-              <Tooltip title={t_i18n('No executable tests are currently set for this entity, these can be set in OpenAEV')} placement="top">
+              <Tooltip title="Pending" placement="top">
                 <Avatar className={classes.iconOverlay} sx={{ bgcolor: 'transparent', width: 24, height: 24 }} style={{ top: 24, left: 23 }}>
                   <span style={{ color: theme.palette.text?.primary, fontSize: 18 }}><HourglassEmpty style={{ fontSize: 23 }} /></span>
                 </Avatar>
@@ -186,7 +183,6 @@ const SecurityCoverageScores: FunctionComponent<SecurityCoverageScoresProps> = (
           <div key={coverageResult.coverage_name} className={classes.chartItem}>
             <div className={classes.chartContainer}>
               <div className={classes.chart}>
-                <>test 2</>
                 <Chart options={options} series={series} type="donut" width={70} height={70} />
                 <Tooltip title={coverageResult.coverage_name} placement="top">
                   <Avatar className={classes.iconOverlay} sx={{ bgcolor: 'transparent', width: 25, height: 30 }}>
